+++ cmd/gophermart/server/storage/models/user/balance/withdraw.go
@@ -2,7 +2,7 @@ package balance
 
 // Withdraw описывает структуру списания баллов
 type Withdraw struct {
-	OrderNumber string `json:"order"`                  // Заказ
-	Sum         int    `json:"sum"`                    // Сумма
-	ProcessedAt string `json:"processed_at,omitempty"` // Время
+	OrderNumber string  `json:"order"`                  // Заказ
+	Sum         float32 `json:"sum"`                    // Сумма
+	ProcessedAt string  `json:"processed_at,omitempty"` // Время
 }
